Time: 0 ms (100.00%), Space: 40.4 MB (33.05%) - LeetHub

+++ 2582-pass-the-pillow/2582-pass-the-pillow.java
@@ -3,7 +3,7 @@ class Solution {
         short i = 1;
         short s = 1;
         while ( time > 0 ) {
-            int distance = Math.min(time,n-1);
+            short distance = (short) Math.min(time,n-1);
             i += distance * s;
             time -= distance;
             s *= -1;
